requirements.txt
# Jungeui Lab- API & Scripts (Python 3.10+)
# API server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=44.0.0

# Auth & env
python-dotenv>=1.0.0
bcrypt>=4.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# File upload / utils
python-multipart>=0.0.12

# Scripts (optional)
# pandas>=2.0.0
